+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This is the changelog for [Authress Login](readme.md).
 ## 1.3 ##
 * Automatically trigger credentials checking on load.
 * Support 127.0.0.1 for localhost
+* Add the `updateExtensionAuthenticationRequest` method to support handling platform extension login.
 
 ## 1.2 ##
 * Use builtin `crypto.subtle` for all crypto operations.